Implementation of *MappingBuilder and more free standing functions
Implement most of the MappingBuilder and ScenarioMappingBuilder functions.
Add more free standing functions for the WireMock DSL.

+++ src/http/delay_distribution.rs
@@ -10,7 +10,7 @@ pub enum DelayDistribution {
     },
     #[serde(rename = "uniform")]
     Uniform {
-        lower: u16,
-        upper: u16,
+        lower: u32,
+        upper: u32,
     },
 }
